✨ Add SpecifiedAttribute and RequiredValueAttribute to OptionalValues.DataAnnotations. (#17)
Add `SpecifiedAttribute` and `RequiredValueAttribute` to `OptionalValues.DataAnnotations`.

`[Specified]` will only validate that the value is Specified.
`[RequiredValue]` will validate that the value is Specified, not-null and not an empty string.

```csharp
// Must be present; can be null
[Specified]
public OptionalValue<string?> DisplayName { get; set; }

// Must be present and have a value ("" allowed here)
[RequiredValue(AllowEmptyStrings = true)]
public OptionalValue<string?> Nickname { get; set; }
```

--

* 🔖 Bump version for adding Optional Required validation checks

* ✨ Add OptionalRequiredAttribute

Add OptionalRequiredAttribute to OptionalValues.DataAnnotations which will give an error when the value is Unspecified, null, or is an empty string. It mimics the behavior of the required attribute.

* 🚨 Fix warnings in test because of incompatibility of custom converters with source generators

* ✅ Update OpenApi examples to include OptionalRequired.

This is to verify that it marks those properties as required.

* ✨ Add OptionalSpecifiedAttribute

The OptionalSpecifiedAttribute validates that a property is Specified.

* ♻️ Rename OptionalSpecified to Specified and OptionalRequired to RequiredValue

* ✨ Handle [Specified] in OpenApi

* 📝 Update documentation